added -Wall and cleaned up racecheck_unittest.cc
git-svn-id: http://data-race-test.googlecode.com/svn/trunk@1010 93a3164d-3a44-0410-8ae5-4baa038b0729

+++ unittest/racecheck_unittest.cc
@@ -1797,7 +1797,7 @@ void Run() {
   for (int i = 0; i < N_free; i++) delete ARR[i];
   delete [] ARR;
   
-  for (int i = 0; i < mus.size(); i++) {
+  for (size_t i = 0; i < mus.size(); i++) {
     delete mus[i];
   }
 }
